frontend/requirements.txt
streamlit==1.34.0
langchain==0.1.14
langchain-google-genai==0.0.4
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.14
langchain-google-genai==0.0.4
google-generativeai==0.3.1
streamlit==1.34.0
python-dotenv
